🎨 Move useTaskStore to store dir
moves useTaskStore to store directory
